[Server] Update MyMenu & Add test

+++ server/src/main/java/com/ewha/heydongdong/module/model/dto/MyMenuDto.java
@@ -14,7 +14,7 @@ import java.sql.Timestamp;
 @ToString
 public class MyMenuDto {
 
-    private Integer myMenuId;
+    private Long myMenuId;
     private MenuInMyMenuDto menuInOrder;
     @JsonFormat(shape = JsonFormat.Shape.STRING, pattern = "yyyy-MM-dd hh:mm:ss")
     private Timestamp addAt;
